refactor: modularize stripe handlers
