+++ functions/src/index.ts
@@ -22,21 +22,7 @@ export {
   updateUserProfileCallable,
   completeSignupAndProfile,
 } from './features/users';
-export {
-  startSubscriptionCheckout,
-  startOneTimeTokenCheckout,
-  startTokenCheckout,
-  createCheckoutSession,
-  createStripeSubscriptionIntent,
-  startDonationCheckout,
-  startCheckoutSession,
-  createStripeCheckout,
-  createStripeSetupIntent,
-  finalizePaymentIntent,
-  createTokenPurchaseSheet,
-  createSubscriptionSession,
-} from './stripeHandlers';
+export * from './stripe';
 
 export { onCompletedChallengeCreate } from './firestoreArchitecture';
-export { handleStripeWebhookV2 } from './stripeWebhooks';
 export { cleanLegacySubscriptionFields } from './cleanLegacySubscriptionFields';
